+++ modules/modelSetup/BaseHunyuanVideoSetup.py
@@ -24,6 +24,12 @@ from modules.util.TrainProgress import TrainProgress
 import torch
 from torch import Tensor
 
+PRESETS = {
+    "attn-mlp": ["attn", "ff.net"],
+    "attn-only": ["attn"],
+    "full": [],
+}
+
 
 class BaseHunyuanVideoSetup(
     BaseModelSetup,
