+++ modules/modelSetup/BasePixArtAlphaSetup.py
@@ -23,6 +23,11 @@ from modules.util.TrainProgress import TrainProgress
 import torch
 from torch import Tensor
 
+PRESETS = {
+    "attn-mlp": ["attn1", "attn2", "ff.net"],
+    "attn-only": ["attn1", "attn2"],
+    "full": [],
+}
 
 class BasePixArtAlphaSetup(
     BaseModelSetup,
