+++ modules/modelSetup/HunyuanVideoLoRASetup.py
@@ -11,12 +11,6 @@ from modules.util.TrainProgress import TrainProgress
 
 import torch
 
-PRESETS = {
-    "attn-mlp": ["attn", "ff.net"],
-    "attn-only": ["attn"],
-    "full": [],
-}
-
 
 class HunyuanVideoLoRASetup(
     BaseHunyuanVideoSetup,
@@ -96,7 +90,7 @@ class HunyuanVideoLoRASetup(
             ) if create_te2 else None
 
         model.transformer_lora = LoRAModuleWrapper(
-            model.transformer, "lora_transformer", config, config.lora_layers.split(",")
+            model.transformer, "lora_transformer", config, config.layer_filter.split(",")
         )
 
         if model.lora_state_dict:
