+++ modules/modelSetup/PixArtAlphaLoRASetup.py
@@ -9,12 +9,6 @@ from modules.util.TrainProgress import TrainProgress
 
 import torch
 
-PRESETS = {
-    "attn-mlp": ["attn1", "attn2", "ff.net"],
-    "attn-only": ["attn1", "attn2"],
-    "full": [],
-}
-
 
 class PixArtAlphaLoRASetup(
     BasePixArtAlphaSetup,
@@ -77,7 +71,7 @@ class PixArtAlphaLoRASetup(
         ) if create_te else None
 
         model.transformer_lora = LoRAModuleWrapper(
-            model.transformer, "lora_transformer", config, config.lora_layers.split(",")
+            model.transformer, "lora_transformer", config, config.layer_filter.split(",")
         )
 
         if model.lora_state_dict:
