+++ modules/modelSetup/StableDiffusionXLLoRASetup.py
@@ -9,12 +9,6 @@ from modules.util.TrainProgress import TrainProgress
 
 import torch
 
-PRESETS = {
-    "attn-mlp": ["attentions"],
-    "attn-only": ["attn"],
-    "full": [],
-}
-
 
 class StableDiffusionXLLoRASetup(
     BaseStableDiffusionXLSetup,
@@ -90,7 +84,7 @@ class StableDiffusionXLLoRASetup(
         ) if create_te2 else None
 
         model.unet_lora = LoRAModuleWrapper(
-            model.unet, "lora_unet", config, config.lora_layers.split(",")
+            model.unet, "lora_unet", config, config.layer_filter.split(",")
         )
 
         if model.lora_state_dict:
