+++ modules/modelSetup/WuerstchenFineTuneSetup.py
@@ -1,6 +1,7 @@
 from modules.model.WuerstchenModel import WuerstchenModel
 from modules.modelSetup.BaseWuerstchenSetup import BaseWuerstchenSetup
 from modules.util.config.TrainConfig import TrainConfig
+from modules.util.ModuleFilter import ModuleFilter
 from modules.util.NamedParameterGroup import NamedParameterGroupCollection
 from modules.util.optimizer_util import init_model_parameters
 from modules.util.TrainProgress import TrainProgress
@@ -38,7 +39,8 @@ class WuerstchenFineTuneSetup(
                 "prior_embeddings"
             )
 
-        self._create_model_part_parameters(parameter_group_collection, "prior_prior", model.prior_prior, config.prior)
+        self._create_model_part_parameters(parameter_group_collection, "prior_prior", model.prior_prior, config.prior,
+                                           freeze=ModuleFilter.create(config), debug=config.debug_mode)
 
         return parameter_group_collection
