+++ training_presets/#chroma Finetune 16GB.json
@@ -42,5 +42,7 @@
             "warmup_init": false,
             "weight_decay": 0.0
         }
-    }
+    },
+    "layer_filter": "transformer_block",
+    "layer_filter_preset": "blocks"
 }
