+++ training_presets/#chroma Finetune 24GB.json
@@ -39,5 +39,7 @@
             "warmup_init": false,
             "weight_decay": 0.0
         }
-    }
+    },
+    "layer_filter": "transformer_block",
+    "layer_filter_preset": "blocks"
 }
