+++ training_presets/#chroma LoRA 16GB.json
@@ -21,6 +21,6 @@
     "output_dtype": "BFLOAT_16",
     "timestep_distribution": "INVERTED_PARABOLA",
     "noising_weight": 7.7,
-    "lora_layers": "attn,ff.net",
-    "lora_layer_preset": "attn-mlp"
+    "layer_filter": "attn,ff.net",
+    "layer_filter_preset": "attn-mlp"
 }
